% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_evaluation.R
\name{campaignevaluation}
\alias{campaignevaluation}
\title{SGB with user-defined loss function}
\usage{
campaignevaluation(
  y,
  treated,
  scores,
  m,
  delta,
  conditional = TRUE,
  opttargetperc,
  increment = 0.1,
  plot
)
}
\arguments{
\item{y}{post-intervention churn}

\item{treated}{whether the customer is treated}

\item{scores}{estimated scores}

\item{m}{post-intervention cash flow}

\item{delta}{intervention cost}

\item{conditional}{whether the offer is conditional on renewal}

\item{opttargetperc}{vector of length k of given target size for each method}

\item{increment}{share of customers added to target size, if "unit",
customers are added one by one}

\item{plot}{whether the campaign profit curve should be plotted}
}
\value{
A list containing holdout campaign profit at optimal target size
(as determined from the validation sample), campaign profit curve
on the test sample, range of target sizes, and
the optimal target size determined on the validation sample.

\itemize{
\item campaign.evaluation
\item campaign.profit.curve
\item target.sizes
\item targetsize.maxprofit
}
}
\description{
SGB with user-defined loss function
}
\details{
Calculate the (profit) lift for a given target size
based on the estimated scores
}
