% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_profit_lift.R
\name{campaignprofitlift}
\alias{campaignprofitlift}
\title{Campaign profit lift}
\usage{
campaignprofitlift(y, treated, m, delta, conditional = TRUE)
}
\arguments{
\item{y}{post-intervention churn of targeted customers}

\item{treated}{whether the customer is treated}

\item{m}{post-intervention cash flow}

\item{delta}{intervention cost}

\item{conditional}{whether the offer is conditional
on renewal (default: TRUE)}
}
\value{
campaign profit
}
\description{
Calculate the Campaign Profit Lift - returns NA when one group is empty.
}
