% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_three_samples.R
\name{create.3samples}
\alias{create.3samples}
\title{Randomly divide a data set in three samples}
\usage{
create.3samples(mydata, size.train, size.valid, size.test, myseed = 268237)
}
\arguments{
\item{mydata}{input data to be divided in 3 samples}

\item{size.train}{number of observations in the calibration sample}

\item{size.valid}{number of observations in the validation sample}

\item{size.test}{number of observations in the test sample}

\item{myseed}{set seed for reproducibility}
}
\value{
A list containing campaign profit curve, target sizes,

\itemize{
\item data.train
\item data.valid
\item data.test
\item train.index
\item test.index
}
}
\description{
Randomly divide a data set in three samples.
}
