% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_size_campaign_evaluation.R
\name{fixedsizecampaignevaluation}
\alias{fixedsizecampaignevaluation}
\title{Fixed target size determination}
\usage{
fixedsizecampaignevaluation(
  y,
  treated,
  scores,
  m,
  delta,
  conditional = TRUE,
  fixedsizeperc,
  increment = 0.1
)
}
\arguments{
\item{y}{post-intervention churn}

\item{treated}{whether the customer is treated}

\item{scores}{estimated scores}

\item{m}{post-intervention cash flow}

\item{delta}{intervention cost}

\item{conditional}{whether the offer is conditional on renewal (default:TRUE)}

\item{fixedsizeperc}{target size in percentage customers being targeted}

\item{increment}{share of customers added to target size, if "unit",
customers are added one by one}
}
\value{
A list containing campaign profit and target size

\itemize{
\item profit.fixedsize
\item targetsize.fixedsize
}
}
\description{
Specifies the target size exogenously and calculate the corresponding
campaign profit.
}
