% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{gini}
\alias{gini}
\title{Gini}
\usage{
gini(y, p)
}
\arguments{
\item{y}{a vector of length n containing the actual churn outcome of
all customers}

\item{p}{a vector of length n containing the estimated probability or
score to churn of all customers}
}
\value{
the gini coefficient
}
\description{
This function calculates the gini coefficient of a churn model,
given the estimated probabilities to churn.
}
