% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_lift_model.R
\name{lift.procedure}
\alias{lift.procedure}
\title{Lift Model Procedure}
\usage{
lift.procedure(myseed, mydata, colnumbers.covariates, delta, conditional)
}
\arguments{
\item{myseed}{seed to split the data set in 3 samples}

\item{mydata}{select the data to be split in 3 samples}

\item{colnumbers.covariates}{column numbers of predictors excluding w}

\item{delta}{cost of intervention}

\item{conditional)}{whether the offer is conditional or not}
}
\value{
List containing seed (myseed),
whether the offer is conditional or not (conditional), and
cost of intervention (delta).
Followed by
data (train.data, valid.data and test.data),
probability of churn (non-targeted r0 /targeted r1),
expected margin (non-targeted m0 /targeted m1),
and lift (rmlift) for respectively
train for train, valid, and test samples.

\itemize{
\item myseed
\item conditional
\item delta
\item train.data
\item valid.data
\item test.data
\item r0.train
\item r0.valid
\item r0.test
\item r1.train
\item r1.valid
\item r1.test
\item m0.train
\item m0.valid
\item m0.test
\item m1.train
\item m1.valid
\item m1.test
\item rmlift.train
\item rmlift.valid
\item rmlift.test
}
}
\description{
Procedure to Lift Models
}
