% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mysynthdata}
\alias{mysynthdata}
\title{Mysynthdata}
\format{
A data frame with 200 rows and 14 variables:
\describe{
\item{\code{w}}{integer the treatment dummy}
\item{\code{x1}}{integer covariate 1}
\item{\code{x2}}{integer covariate 2}
\item{\code{x3}}{integer covariate 3}
\item{\code{x4}}{integer covariate 4}
\item{\code{x5}}{integer covariate 5}
\item{\code{x6}}{integer covariate 6}
\item{\code{x7}}{integer covariate 7}
\item{\code{x8}}{integer covariate 8}
\item{\code{x9}}{integer covariate 9}
\item{\code{x10}}{integer covariate 10}
\item{\code{x11}}{integer covariate 11}
\item{\code{y}}{integer post-intervention churn (0/1)}
\item{\code{revenues}}{double post-intervention cash flows
("m" in the "Managing Churn to Maximize Profits"
manuscript)}
}
}
\usage{
mysynthdata
}
\description{
Synthetic data set for Managing Churn to Maximize Profits
}
\keyword{datasets}
