% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_per_decile.R
\name{overlap.per.decile}
\alias{overlap.per.decile}
\title{Customer overlap}
\usage{
overlap.per.decile(score1, score2, splitn = 10, plot = TRUE)
}
\arguments{
\item{score1}{estimated score of first method for which to calculate overlap}

\item{score2}{estimated score of second method for which to calculate overlap}

\item{splitn}{number of deciles (default: 10)}

\item{plot}{whether the overlap curve should be plotted (default: TRUE)}
}
\value{
Percentage Customers Overlapping.
}
\description{
Calculate customer overlap between two rankings
}
\details{
Overlap figure:
\enumerate{
\item rank customers according to all methods' scores
\item split in decile
\item calculate per decile, the \% customers overlapping in that decile
}

See also Figure 4 in Lemmens, Aurélie, and Sunil Gupta.
"Managing churn to maximize profits." Marketing Science, (2020).
}
