% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgb_user_defined_loss.R
\name{sgb.loss}
\alias{sgb.loss}
\title{SGB with user-defined loss function}
\usage{
sgb.loss(
  data.train,
  data.valid,
  data.test,
  otherdata1,
  otherdata2,
  otherdata3,
  otherdata4,
  otherdata5,
  otherdata6,
  y.type,
  pos.y,
  pos.covariates,
  pi,
  pi.valid,
  loss,
  reorder,
  conditional,
  miniter,
  maxiter,
  rho,
  stoch,
  ratio,
  m.convergence.train,
  m.convergence.valid,
  m.convergence.test,
  delta.convergence,
  increment.convergence,
  m0.train,
  m1.train,
  m0.valid,
  m1.valid,
  m0.test,
  m1.test,
  verbose = TRUE
)
}
\arguments{
\item{data.train}{training data set}

\item{data.valid}{validation data set}

\item{data.test}{test data set}

\item{otherdata1}{other data used for the reordered classic loss, 1}

\item{otherdata2}{other data used for the reordered classic loss, 2}

\item{otherdata3}{other data used for the reordered classic loss, 3}

\item{otherdata4}{other data used for the reordered classic loss, 4}

\item{otherdata5}{other data used for the reordered classic loss, 5}

\item{otherdata6}{other data used for the reordered classic loss, 6}

\item{y.type}{churn variable,  and}

\item{pos.y}{positions (col) of churn}

\item{pos.covariates}{covariates}

\item{pi}{individual expected profit lifts used as DV and weights
in the profit loss function}

\item{pi.valid}{individual expected profit lifts used as DV and weights
in the profit loss function}

\item{loss}{specify loss function and type of offer}

\item{reorder}{whether scores should be reordered}

\item{conditional}{whether the offer is conditional on renewal}

\item{miniter}{miniter hyperparameter}

\item{maxiter}{maxiter hyperparameter}

\item{rho}{learning rate}

\item{stoch}{use stochastic GB or regular GB}

\item{ratio}{fraction of observations to be sampled in case of stochastic GB}

\item{m.convergence.train}{training scores to determine convergence by
calculating optimal campaign profit at each
iteration}

\item{m.convergence.valid}{validation scores to determine convergence
by calculating optimal campaign profit at each
iteration}

\item{m.convergence.test}{test scores to determine convergence by
calculating optimal campaign profit at each
iteration}

\item{delta.convergence}{intervention cost to determine convergence by
calculating optimal campaign profit at each
iteration}

\item{increment.convergence}{granularity level to determine convergence by
calculating optimal campaign profit at each
iteration}

\item{m0.train}{non-targeted margin scores reordered classic
loss train sample}

\item{m1.train}{targeted margin scores reordered classic loss train sample}

\item{m0.valid}{non-targeted margin scores reordered classic loss
validation sample}

\item{m1.valid}{targeted margin scores reordered classic loss
validation sample}

\item{m0.test}{non-targeted margin scores reordered classic loss test sample}

\item{m1.test}{targeted margin scores reordered classic loss test sample}

\item{verbose}{verbose output (default: TRUE)}
}
\value{
A list containing best iteration of sgb (best.iteration.valid),
scores at best iteration (fmts),
variables' importance (Variable.Importance),
campaign profit at best target size for each iteration of SGB
(maxlift.convergence.valid),
target size at best iteration optimized on valid sample
(targetsize.convergence.valid),
and other SGB outcome measures.

\itemize{
\item iterations
\item best.iteration.valid
\item LogLik
\item ytrain
\item yv
\item yts
\item fmt
\item fmv
\item fmts
\item Variable.Importance
\item maxlift.convergence.valid
\item targetsize.convergence.valid
}
}
\description{
SGB with user-defined loss function.
}
\details{
This function applies the SGB algorithm with a user-defined loss function.
It calibrates the model on a calibration sample and generates predictions
on a validation and test samples.
}
