% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetsize_optimization.R
\name{targetsizeoptimization}
\alias{targetsizeoptimization}
\title{Target Size Optimization}
\usage{
targetsizeoptimization(
  y,
  treated,
  scores,
  m,
  delta,
  conditional = TRUE,
  increment = 0.1,
  plot = TRUE
)
}
\arguments{
\item{y}{post-intervention churn dummy}

\item{treated}{treatment dummy}

\item{scores}{predictions}

\item{m}{post-intervention cash flows}

\item{delta}{intervention cost}

\item{conditional}{conditional offer or not (default: TRUE)}

\item{increment}{share of customers added at a time to build the campaign
profit curve (default: .1)}

\item{plot}{plot the campaign profit curve (default: TRUE)}
}
\value{
A list containing campaign profit curve, target sizes,
maximum profit and related variables.

\itemize{
\item campaign.profit.curve
\item target.sizes
\item maxprofit
\item targetsize.maxprofit
\item maxprofit.trimmed
\item targetsize.maxprofit.trimmed
\item secondbesttargetsize.trimmed
\item increment
}
}
\description{
Target Size Optimization in RCT.
}
