% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_wsgb_model.R
\name{wsgb.procedure}
\alias{wsgb.procedure}
\title{W-SGB}
\usage{
wsgb.procedure(
  myseed,
  mydata,
  pos.covariates,
  pos.y,
  delta,
  conditional,
  loss,
  reorder = FALSE,
  pi,
  pi.valid,
  increment,
  maxiter = 3000,
  miniter = 1500,
  rho = 0.001,
  stoch = TRUE,
  ratio = 0.3,
  m0.train,
  m1.train,
  m0.valid,
  m1.valid,
  m0.test,
  m1.test,
  verbose = TRUE
)
}
\arguments{
\item{myseed}{seed for reproducibility of the results}

\item{mydata}{input data}

\item{pos.covariates}{position (column number of the covariates)}

\item{pos.y}{indicates the position (column number) of y in the data set}

\item{delta}{intervention cost}

\item{conditional}{whether the offer is conditional on renewal}

\item{loss}{loss function}

\item{reorder}{whether scores should be reordered (default: FALSE)}

\item{pi}{individual expected profit lifts used as
DV and weights in the profit loss function - not needed for
classic loss increment, used to assess convergence in SGB
when computing optimal target size}

\item{pi.valid}{individual expected profit lifts used as
DV and weights in the profit loss function - not needed for
classic loss increment, used to assess convergence in SGB
when computing optimal target size}

\item{maxiter}{maxiter hyperparameter SGB (default: 3000)}

\item{miniter}{miniter hyperparameter SGB (default: 1500)}

\item{rho}{rho hyperparameter SGB     (default: .001)}

\item{stoch}{stoch hyperparameter SGB   (default: TRUE)}

\item{ratio}{ratio hyperparameter SGB   (default: .3)}

\item{m0.train}{non-targeted margin scores reordered classic loss train sample}

\item{m1.train}{targeted margin scores reordered classic loss train sample}

\item{m0.valid}{non-targeted margin scores reordered classic loss valid sample}

\item{m1.valid}{targeted margin scores reordered classic loss valid sample}

\item{m0.test}{non-targeted margin scores reordered classic loss test sample}

\item{m1.test}{targeted margin scores reordered classic loss test sample}

\item{verbose}{verbose output (default: TRUE)}
}
\value{
List with W-SGB outcome measures.

\itemize{
\item myseed
\item conditional
\item delta
\item train.data
\item valid.data
\item test.data
\item scores.train
\item scores.valid
\item scores.test
\item best.iteration.valid
\item LogLik
\item Variable.Importance
\item maxlift.convergence.valid
\item targetsize.convergence.valid
}
}
\description{
Procedure W-SGB Model
}
