% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_evaluation.R
\name{campaignevaluation}
\alias{campaignevaluation}
\title{Generate Campaign Profit Curve}
\usage{
campaignevaluation(
  y,
  treated,
  scores,
  m,
  delta,
  conditional = TRUE,
  opttargetperc,
  increment = 0.1,
  plot
)
}
\arguments{
\item{y}{post-intervention churn (binary vector of length n, 0/1)}

\item{treated}{whether the customer is treated (binary vector of length n, 0/1)}

\item{scores}{estimated scores (vector of length n) provided by a model, e.g. profit-based SGB)}

\item{m}{post-intervention cash flow (vector of length n)}

\item{delta}{intervention cost (numeric)}

\item{conditional}{whether the offer is conditional on renewal (logical)}

\item{opttargetperc}{selected target size, in percentage}

\item{increment}{fraction of customers added to target size when generating the profit curve, if "unit",
customers are added one by one}

\item{plot}{whether the campaign profit curve should be plotted (logical)}
}
\value{
A list containing the campaign profit at target size "opttargetperc", the campaign profit curve, the range of target sizes, and
the number of customers targeted (as requested by "opttargetperc").

\itemize{
\item campaign.evaluation: campaign profit at target size "opttargetperc"
\item campaign.profit.curve: campaign profit value for each target size (given by target.sizes)
\item target.sizes: set of target sizes at which the campaign profit curve is calculated
\item targetsize.maxprofit: selected target size (in units, as request by "opttargetperc").
}
}
\description{
This function calculates the campaign profit at a given target size based on the estimated scores. It also returns the campaign profit curve for all possible target sizes (from 0% to 100%)
}
\details{
Calculate the campaign profit at a given target size based on the estimated scores
}
